[rename] #20 s3관련 설정 external로 이동
