+++ src/main/java/org/sopt/sweet/global/external/s3/S3Config.java
@@ -1,4 +1,4 @@
-package org.sopt.sweet.global.config;
+package org.sopt.sweet.global.external.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
